+++ source/extensions/isaacsim.core.prims/python/impl/single_articulation.py
@@ -86,7 +86,6 @@ class SingleArticulation(_SinglePrimWrapper):
         visible: Optional[bool] = None,
         reset_xform_properties: bool = True,
         articulation_controller: Optional["ArticulationController"] = None,
-        enable_residual_reports: bool = False,
     ) -> None:
         from isaacsim.core.simulation_manager import SimulationManager
 
@@ -116,7 +115,6 @@ class SingleArticulation(_SinglePrimWrapper):
             scales=scale,
             visibilities=visible,
             reset_xform_properties=reset_xform_properties,
-            enable_residual_reports=enable_residual_reports,
         )
         self._articulation_controller = articulation_controller
         if self._articulation_controller is None:
@@ -402,38 +400,6 @@ class SingleArticulation(_SinglePrimWrapper):
         """
         velocities = self._articulation_view.get_velocities()
         return velocities[0]
-
-    def get_position_residual(self, report_max: Optional[bool] = True) -> float:
-        """Get physics solver position residuals for articulations. This is the residual across all joints that are part of articulations.
-
-        The solver residuals are computed according to impulse variation normalized by the effective mass.
-
-        Args:
-            report_max (Optional[bool]): whether to report max or RMS residual. Defaults to True, i.e. max criteria
-
-        Returns:
-            float: solver position/velocity max/rms residual.
-        """
-        result = self._articulation_view.get_position_residuals(report_max=report_max)
-        if result is not None:
-            return result[0]
-        return result
-
-    def get_velocity_residual(self, report_max: Optional[bool] = True) -> float:
-        """Get physics solver velocity residuals for articulations. This is the residual across all joints that are part of articulations.
-
-        The solver residuals are computed according to impulse variation normalized by the effective mass.
-
-        Args:
-            report_max (Optional[bool]): whether to report max or RMS residual. Defaults to True, i.e. max criteria
-
-        Returns:
-            float: solver velocity max/rms residual.
-        """
-        result = self._articulation_view.get_velocity_residuals(report_max=report_max)
-        if result is not None:
-            return result[0]
-        return result
 
     def set_joint_positions(
         self, positions: np.ndarray, joint_indices: Optional[Union[List, np.ndarray]] = None