+++ source/extensions/isaacsim.core.api/docs/CHANGELOG.md
@@ -1,4 +1,8 @@
 # Changelog
+## [5.0.0] - 2026-01-20
+### Removed
+- Remove deprecated PhysX residual reporting APIs (`enable_residual_reporting`, `get_solver_velocity_residual`, `get_solver_position_residual`) from `PhysicsContext`.
+
 ## [4.11.5] - 2026-01-06
 ### Changed
 - Migrate more events to Events 2.0.